Refactored Inventory to inherit from IngredientCatalogue such that methods previously in Inventory can now be used in Recipe as well.
